requirements.txt
click>=8.1
rich>=13.7
pytest>=8.0
coverage>=7.6
radon>=6.0
gitpython>=3.1
jinja2>=3.1
lxml>=5.2
networkx>=3.3
# Advanced features:
mutmut>=2.4
z3-solver>=4.13
